refactor : category 컴포넌트 소분
