+++ src/features/list/components/category/utils.ts
@@ -1,0 +1,7 @@
+export function fmtDateLabel(d: Date | null) {
+  if (!d) return "날짜 전체";
+  const y = d.getFullYear();
+  const m = String(d.getMonth() + 1).padStart(2, "0");
+  const day = String(d.getDate()).padStart(2, "0");
+  return `${y}.${m}.${day}`;
+}
